requirements.txt
appdirs==1.4.4
apsw==3.36.0.post1
apturl==0.5.2
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssselect2==0.7.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.12.0
dataclasses==0.6
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
duplicity==0.8.21
eth-hash==0.3.3
eth-typing==2.3.0
eth-utils==1.10.0
eth2spec==0.11.3
fasteners==0.14.1
feedparser==6.0.8
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httplib2==0.20.2
i3ipc==2.2.1
idna==3.3
ifaddr==0.1.7
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lru-dict==1.1.6
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==0.0.0
msgpack==1.0.3
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
proglog==0.1.10
prompt-toolkit==3.0.28
protobuf==3.12.4
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.9.0
ptyprocess==0.7.0
py-ecc==2.0.0
py7zr==0.11.3+dfsg
pycairo==1.20.1
pychm==0.8.6
pycryptodome==3.9.4
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-xapp==2.2.1
python-xlib==0.29
pythondialog==3.5.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2021.11.10
remerkleable==0.1.13
reportlab==3.6.12
repoze.lru==0.7
requests==2.27.1
requests-toolbelt==0.9.1
rofication==1.2.2
Routes==2.5.1
ruamel.yaml==0.16.5
scipy==1.8.0
SecretStorage==3.3.1
setproctitle==1.2.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
svglib==1.4.1
sympy==1.9
systemd-python==234
texttable==1.6.4
tinycss2==1.2.1
tls-client==0.1.5
toolz==0.12.0
tqdm==4.64.1
traitlets==5.1.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
xdg==5
xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
